requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
asgiref==3.8.1
django==5.1
sqlparse==0.5.1
tzdata==2024.1
pyyaml==6.0.2
psycopg2==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-crontab==0.7.1